+++ spine-ios/CHANGELOG.md
@@ -17,3 +17,4 @@
 - Load assets from different sources (#13)
 - Add `Obj-C` + `UIKit` sample (#14)
 - Support CocoaPods (#15)
+- Metal Best Practices (#16)